+++ CHANGELOG.md
@@ -1,10 +1,18 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [v0.0.1]
+
+### Fixed
+
+- The dynamic analyzer/controller class loaders now skip searching the
+  ALICA_interfaces.jar for analyzer or controller classes.
+
 ## [v0.0.0]
 
 ### Added
 
 - Initial project files.
 
+[v0.0.1]: https://github.com/LEB-EPFL/ALICA_interfaces/releases/tag/0.0.1
 [v0.0.0]: https://github.com/LEB-EPFL/ALICA_interfaces/releases/tag/0.0.0